[lazy-defs] Rename definitions_loader.py -> definitions_load_context.py
