+++ python_modules/libraries/dagster-powerbi/dagster_powerbi/resource.py
@@ -13,7 +13,10 @@ from dagster import (
     multi_asset,
 )
 from dagster._config.pythonic_config.resource import ResourceDependency
-from dagster._core.definitions.definitions_loader import DefinitionsLoadContext, DefinitionsLoadType
+from dagster._core.definitions.definitions_load_context import (
+    DefinitionsLoadContext,
+    DefinitionsLoadType,
+)
 from dagster._core.definitions.events import Failure
 from dagster._core.execution.context.asset_execution_context import AssetExecutionContext
 from dagster._utils.cached_method import cached_method
